Fix compile errors: CoinType cast and stale ScoreManager test fields
- GameSceneManager.cs: remove (int) cast in AddCoins call — method expects CoinType enum
- ScoreManagerTest.cs: replace removed whiteCoinsToWin/blackCoinsToWin/queenToWin with coinsToWin

+++ Assets/_Project/Scripts/UI/GameSceneManager.cs
@@ -133,7 +133,7 @@ public class GameSceneManager : MonoBehaviour
         if (scoreManager != null)
         {
             int currentPlayer = turnManager.GetCurrentPlayer();
-            scoreManager.AddCoins(currentPlayer, (int)coinType, 1);
+            scoreManager.AddCoins(currentPlayer, coinType, 1);
         }
     }
